clean of the code main.c

+++ main.c
@@ -7,13 +7,45 @@
 #include <sys/types.h>
 #include <sys/user.h>
 
-long str_to_hexa(char *s)
-{
+/*
+Structure du programme tracé
+
+// Calcule a^b en b multiplications
+int puissance(int a, int b);
+
+// Calcule a^b par exponentiation rapide
+// L'objectif finale est d'écrire nous même
+// cette fonction dans la heap
+int puissance_opti(int a, int b);
+
+int main() {
+  while 1 {
+    init = time();
+    for i in 0..2000 {
+      for j in 0..2000 {
+        puissance(i, j);
+      }
+    }
+    end = time();
+    printf("%d", end - init);
+    // Prend environ 7.5s pour a = 2000 et b = 2000
+    // avec la fonction puissance non optimisé
+  }
+}
+*/
+/// Objectif Final du TP :
+// Lancer puissance_opti au lieu de puissance
+// Cette fonction est écrite dans la heap
+// On va donc écrire dans la heap le code de puissance_opti
+// et modifier la fonction puissance pour qu'elle appelle
+// puissance_opti à la place de puissance
+
+/// Convertit une chaine de caractère s en hexadécimal
+long str_to_hexa(char *s) {
   long result = 0;
   int c;
   int w = 0;
-  while (w < strlen(s) - 1)
-  {
+  while (w < strlen(s) - 1) {
     result = result << 4;
     if (c = (*(s + w) - '0'), (c >= 0 && c <= 9))
       result |= c;
@@ -28,120 +60,116 @@ long str_to_hexa(char *s)
   return result;
 }
 
-// int read_memory(long adr, int length, int pid, char *res)
-//
-int read_memory(long adr, int length, int pid, char *res)
-{
+/// Lis la mémoire d'un processus tracé (pid) à l'adresse adr
+/// - buffer : pointeur vers la zone mémoire où on veut stocker le résultat
+/// - pid : pid du processus tracé
+/// - adr : addresse à laquelle on veut lire
+/// - length : nombre d'octets à lire
+/// > return : EXIT_FAILURE si erreur, EXIT_SUCCESS sinon
+int read_memory(long adr, int length, int pid, char *buffer) {
   char path[16];
   sprintf(path, "/proc/%d/mem", pid);
-
   FILE *f = fopen(path, "r+");
-  if (f == NULL)
-  {
+  if (f == NULL) {
     printf("Erreur dans l'ouverture du fichier memoire du processus\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  if (fseek(f, adr, SEEK_SET) == -1)
-  {
+  if (fseek(f, adr, SEEK_SET) == -1) {
     printf("Erreur de placement dans le fichier\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  if (fread(res, length, sizeof(char), f) == 0)
-  {
+  if (fread(buffer, length, sizeof(char), f) == 0) {
     printf("Erreur de lecture dans le fichier\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  fclose(f);
-  return 0;
+  if (fclose(f) != 0) {
+    printf("Erreur de fermeture du fichier\n");
+    return EXIT_FAILURE;
+  };
+  return EXIT_SUCCESS;
 }
 
-// TODO rename res
-int write_memory(long adr, char *inst, int length, int pid, char *res)
-{
+/// Ecrit dans la mémoire d'un processus tracé (pid) à l'adresse adr
+/// - buffer : pointeur vers la zone mémoire où on veut stocker les lignes de
+///            code effacées.
+/// - pid : pid du processus tracé
+/// - adr : addresse à laquelle on veut lire
+/// - length : nombre d'octets à lire
+/// - inst : instruction à écrire (de taille length)
+/// > return : EXIT_FAILURE si erreur, EXIT_SUCCESS sinon
+int write_memory(long adr, char *inst, int length, int pid, char *buffer) {
   char path[16];
-  printf("length : %d\n", length);
+  // printf("length : %d\n", length);
   sprintf(path, "/proc/%d/mem", pid);
   printf("path : %s\n", path);
   printf("adr : 0x%X\n", adr);
-  for (int i = 0; i < length; i++)
-  {
-    printf("inst : 0x%X\n", inst[i]);
+  for (int i = 0; i < length; i++) {
+    // Affichage de toutes les instructions à écrire
+    // printf("inst : 0x%X\n", inst[i]);
   }
-
   FILE *f = fopen(path, "r+");
-  if (f == NULL)
-  {
+  if (f == NULL) {
     printf("Erreur dans l'ouverture du fichier memoire du processus\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  if (fseek(f, adr, SEEK_SET) == -1)
-  {
+  if (fseek(f, adr, SEEK_SET) == -1) {
     printf("Erreur de placement dans le fichier\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  if (fread(res, length, sizeof(char), f) == 0)
-  {
+  if (fread(buffer, length, sizeof(char), f) == 0) {
     printf("Erreur de lecture dans le fichier\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  for (int i = 0; i < length; i++)
-  {
-    printf("instruction %d sauvegardée : 0x%X\n", i, res[i]);
+  for (int i = 0; i < length; i++) {
+    // Affichage de toutes les instructions sauvegardées
+    printf("Instruction %d sauvegardée : 0x%X\n", i, buffer[i]);
   }
-
-  if (fseek(f, adr, SEEK_SET) == -1)
-  {
+  if (fseek(f, adr, SEEK_SET) == -1) {
     printf("Erreur de placement dans le fichier\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-  int test = fwrite(inst, length, sizeof(char), f);
-  if (test == 0)
-  {
+  if (fwrite(inst, length, sizeof(char), f) == 0) {
     printf("Erreur d'ecriture dans le fichier\n");
-    printf("test : %d\n", test);
-    return -1;
+    return EXIT_FAILURE;
   }
-
-  fclose(f);
-  return 0;
+  if (fclose(f) != 0) {
+    printf("Erreur de fermeture du fichier\n");
+    return EXIT_FAILURE;
+  }
+  return EXIT_SUCCESS;
 }
 
-int get_pid()
-{
+/// Récupère le pid du processus à tracer
+/// - return : pid du processus à tracer si succès, EXIT_FAILURE sinon
+int get_pid() {
   system("ps -aux | grep \"./toy_c_program/toy_c_program\" | cut -c 12-17 | "
          "head -n 1 > pid.txt");
-
   FILE *f = fopen("pid.txt", "r");
-  if (f == NULL)
-  {
+  if (f == NULL) {
     printf("Erreur lors de la creation du fichier du pid\n");
-    return -1;
+    return EXIT_FAILURE;
   }
-
   char pid[100];
-
-  fgets(pid, 100, f);
-  fclose(f);
-
+  if (fgets(pid, 100, f) == NULL) {
+    printf("Erreur lors de la lecture du fichier du pid\n");
+    return EXIT_FAILURE;
+  }
+  if (fclose(f) != 0) {
+    printf("Erreur de fermeture du fichier\n");
+    return EXIT_FAILURE;
+  }
   system("rm pid.txt");
-
   printf("pid:%s", pid);
-
-  int i = atoi(pid);
-  printf("atoi(pid):%d\n", i);
-
-  return i;
+  printf("atoi(pid):%d\n", atoi(pid));
+  return atoi(pid);
 }
 
-long get_adr_fun(char *s)
-{
+/// Récupère l'adresse de la fonction à modifier
+/// - s : nom de la fonction
+/// - return : adresse de la fonction si succès, EXIT_FAILURE sinon
+long get_adr_fun(char *s) {
   char sys_call[200];
+  // On écrit dans un fichier temporaire l'adresse de la fonction
   sprintf(
       sys_call,
       "nm toy_c_program/toy_c_program | grep \"%s\" | cut -c 1-16 > adr.txt",
@@ -150,40 +178,52 @@ long get_adr_fun(char *s)
   system(sys_call);
 
   FILE *f = fopen("adr.txt", "r");
-  if (f == NULL)
-  {
+  if (f == NULL) {
     printf("Erreur lors de la creation du fichier de l'adresse\n");
-    return -1;
+    return EXIT_FAILURE;
   }
 
   char adr[100];
-
-  fgets(adr, 100, f);
-  fclose(f);
-
+  if (fgets(adr, 100, f) == NULL) {
+    printf("Erreur lors de la lecture du fichier du pid\n");
+    return EXIT_FAILURE;
+  }
+  if (fclose(f) != 0) {
+    printf("Erreur de fermeture du fichier\n");
+    return EXIT_FAILURE;
+  }
   system("rm adr.txt");
-
   printf("adr:%s", adr);
-  long i = str_to_hexa(adr);
-  printf("hexa adr:%d\n", i);
-
-  return i;
+  printf("hexa adr:%d\n", str_to_hexa(adr));
+  return str_to_hexa(adr);
 }
 
+/// Fonction qui va modifier la fonction puissance pour qu'elle appelle
+/// puissance_opti à la place de puissance
+/// - addr_write_in_fun : adresse de la fonction à modifier
+/// - process_pid : pid du processus à tracer
+/// - addr_fun_call : adresse de la fonction à appeler
+/// - args_fun : arguments de la fonction à appeler
+/// - args_ptr : indique si les arguments sont des pointeurs
+/// - new_value_stack : valeur à mettre dans la pile pour les valeurs
+///                     pointées
+/// - return_param : valeur de retour de la fonction appelée
+/// - stack_push_size : taille de la pile à modifier (pour les paramètres
+/// pointeurs)
+/// > return : EXIT_FAILURE si erreur, EXIT_SUCCESS sinon
 int trap_call_trap(long addr_write_in_fun, int process_pid, long addr_fun_call,
-                   long long int *args_fun, bool *args_ptr, char *new_value_stack,
-                   char *return_param, int stack_push_size)
-{
+                   long long int *args_fun, bool *args_ptr,
+                   char *new_value_stack, char *return_param,
+                   int stack_push_size) {
   int status;
-  // 3. Modifier dynamiquement le code : trap call trap
+  // Ecriture Trap Call Trap au début de la fonction puissance
   const int length = 4;
   char new_inst[4] = {0xCC, 0xFF, 0xD0, 0xCC};
   char sauv_line[4]; // 20 = taille max d'une instruction (x86 CrInGe)
   if (write_memory(addr_write_in_fun, new_inst, length, process_pid,
-                   sauv_line) == -1)
-  {
+                   sauv_line) == EXIT_FAILURE) {
     printf("Erreur lors de l'appel de write_memory pour trap call trap\n");
-    return -1;
+    return EXIT_FAILURE;
   }
 
   ptrace(PTRACE_CONT, process_pid, 0, 0);
@@ -191,8 +231,7 @@ int trap_call_trap(long addr_write_in_fun, int process_pid, long addr_fun_call,
   // Wait for the first trap
   printf("pid: %d first trap\n", waitpid(process_pid, &status, 0));
 
-  // Recupere la valeur des registres
-
+  // Recuperation de la valeur des registres
   struct user_regs_struct regs;
   struct user_regs_struct sauv_regs;
   ptrace(PTRACE_GETREGS, process_pid, 0, &sauv_regs);
@@ -200,42 +239,32 @@ int trap_call_trap(long addr_write_in_fun, int process_pid, long addr_fun_call,
 
   long long int *registr[3] = {&(regs.rdi), &(regs.rsi), &(regs.rdx)};
   int offset = 0;
-  for (int i = 0; i < 3; i++)
-  {
+  for (int i = 0; i < 3; i++) {
     // Modifier les registres pour call la bonne fonction
-    if (args_ptr[i])
-    { // si le premier argument est un pointeur
+    if (args_ptr[i]) { // si le premier argument est un pointeur
       // On incremente le pointeur de pile
       regs.rsp -= args_fun[i];
       // On ajoute une valeur dans la pile
       // printf("regs.rsp : %p\n", regs.rsp);
       // char new_val[4] = {42, 0, 0, 0};
-      char buffer[4] = {0, 0, 0, 0}; // inutile
+      char buffer[4] = {0, 0, 0, 0};
 
       printf("Ecriture stack args_fun %d %d, offset %d\n", i, args_fun[i],
              offset);
-
       if (write_memory(regs.rsp, new_value_stack + offset, args_fun[i],
-                       process_pid, buffer) == -1)
-      {
+                       process_pid, buffer) == -1) {
         printf(
             "Erreur lors de l'appel de write_memory pour changer la stack\n");
-        return -1;
+        return EXIT_FAILURE;
       }
       *(registr[i]) = regs.rsp; // On pases en parametre un pointeur vers le
                                 // sommet de la pile
       offset += args_fun[i];
-    }
-    else
-    {
+    } else {
       *(registr[i]) = args_fun[i];
     }
   }
-
   regs.rax = addr_fun_call;
-
-  // regs.rsi = 0;
-  // regs.rdx = 0;
 
   // Modifier les valeurs des registres
   ptrace(PTRACE_SETREGS, process_pid, 0, &regs);
@@ -247,26 +276,25 @@ int trap_call_trap(long addr_write_in_fun, int process_pid, long addr_fun_call,
 
   ptrace(PTRACE_GETREGS, process_pid, 0, &regs);
   printf("rax : %p\n", regs.rax);
-  for (int i = 0; i < stack_push_size; i++)
-  {
+  for (int i = 0; i < stack_push_size; i++) {
     char res;
-    read_memory(regs.rsp + i * sizeof(char), 1, process_pid, &res);
+    if (read_memory(regs.rsp + i * sizeof(char), 1, process_pid, &res) == EXIT_FAILURE) {
+      printf("Erreur lors de l'appel de read_memory pour lire la stack\n");
+      return EXIT_FAILURE;
+    };
     return_param[i] = res;
   }
 
   // Reecriture des lignes effacees
-  // fwrite sauv_line addr : addr_puissance
-
   char buffer[sizeof(
-      long long int)]; // inutile car stocke trap call trap (ancienne ligne)
+      long long int)]; // stocke les anciennes valeurs des instructions (trap call trap)
 
   // printf("Réecriture des lignes\n");
 
   if (write_memory(addr_write_in_fun, sauv_line, length, process_pid, buffer) ==
-      -1)
-  {
+      EXIT_FAILURE) {
     printf("Erreur lors de l'appel de write_memory pour restore\n");
-    return -1;
+    return EXIT_FAILURE;
   }
   sauv_regs.rip = sauv_regs.rip - 1;
 
@@ -274,11 +302,11 @@ int trap_call_trap(long addr_write_in_fun, int process_pid, long addr_fun_call,
 
   ptrace(PTRACE_CONT, process_pid, 0, 0);
   printf("End of tct\n");
-  return 0;
+  return EXIT_SUCCESS;
 }
 
-int main(int argc, char **argv)
-{
+
+int main(int argc, char **argv) {
   printf("sizeof(int) = %d\n", sizeof(int));
   int process_pid = get_pid();
   long addr_puissance = get_adr_fun(argv[1]);
@@ -289,10 +317,9 @@ int main(int argc, char **argv)
   printf("long addr_mprotect : %p", addr_mprotect);
 
   // 1. Attacher au processus cible
-  if (ptrace(PTRACE_ATTACH, process_pid, 0, 0) == -1)
-  {
+  if (ptrace(PTRACE_ATTACH, process_pid, 0, 0) == -1) {
     printf("ptrace_attach a echoue\n");
-    return -1;
+    return EXIT_FAILURE;
   }
   // {
   //   printf("--------------trap call trap addition\n");
@@ -340,10 +367,8 @@ int main(int argc, char **argv)
   // int args_fun[3] = {sizeof(int), 0, 0};
   // bool args_ptr[3] = {true, false, false};
   int stack_push_size = 0;
-  for (int i = 0; i < 3; i++)
-  {
-    if (args_ptr[i])
-    {
+  for (int i = 0; i < 3; i++) {
+    if (args_ptr[i]) {
       stack_push_size += args_fun[i];
     }
   }
@@ -353,10 +378,9 @@ int main(int argc, char **argv)
   // Tester le cas pathologique ou PC = adr ou PC = adr+1 ou Pc = adr+2
   if (trap_call_trap(addr_puissance, process_pid, addr_posix_memalign, args_fun,
                      args_ptr, new_value_stack, return_stack,
-                     stack_push_size) == -1)
-  {
+                     stack_push_size) == EXIT_FAILURE) {
     printf("Erreur dans trap call trap posix mem align\n");
-    return -1;
+    return EXIT_FAILURE;
   };
 
   long long int *addr_puissance_opti_heap = return_stack;
@@ -365,7 +389,7 @@ int main(int argc, char **argv)
 
   printf("--------------trap call trap mprotect\n");
 
-  char return_stack_mprotect[4] = {0};
+  char return_stack_mprotect[4] = {0, 0, 0, 0};
 
   printf("sizeof ptr: %d\n", sizeof(addr_puissance_opti));
   status;
@@ -377,14 +401,12 @@ int main(int argc, char **argv)
   // int args_fun[3] = {sizeof(int), 0, 0};
   // bool args_ptr[3] = {true, false, false};
   stack_push_size = 0;
-  for (int i = 0; i < 3; i++)
-  {
-    if (args_ptr[i])
-    {
+  for (int i = 0; i < 3; i++) {
+    if (args_ptr[i]) {
       stack_push_size += args_fun[i];
     }
   }
-  char new_value_stack_mprotect[8] = {0};
+  char new_value_stack_mprotect[8] = {0, 0, 0, 0, 0, 0, 0, 0};
   // char new_value_stack_mprotect[8] = {
   //     return_stack[0], return_stack[1], return_stack[2], return_stack[3],
   //     return_stack[4], return_stack[5], return_stack[6], return_stack[7]};
@@ -392,25 +414,23 @@ int main(int argc, char **argv)
   // Tester le cas pathologique ou PC = adr ou PC = adr+1 ou Pc = adr+2
   if (trap_call_trap(addr_puissance, process_pid, addr_mprotect, args_fun,
                      args_ptr, new_value_stack_mprotect, return_stack_mprotect,
-                     stack_push_size) == -1)
-  {
+                     stack_push_size) == EXIT_FAILURE) {
     printf("Erreur dans trap call trap mprotect\n");
-    return -1;
+    return EXIT_FAILURE;
   };
 
   // Ecriture du programme dans la heap
   char buffer[113];
   printf("Ecriture du programme dans la heap\n");
   if (write_memory(addr_puissance_opti_heap, code, size_code, process_pid,
-                   buffer) == -1)
-  {
+                   buffer) == EXIT_FAILURE) {
     printf(
         "Erreur lors de l'appel de write_memory du programme dans la heap\n");
-    return -1;
+    return EXIT_FAILURE;
   }
 
-  // Write in memory Jump to this programm (in the heap) in the function
-  // puissance
+  // Write in memory Jump to this programm (in the heap)
+  // in the function puissance that is called very often
   printf("Write Jump in puissance\n");
   char buffer_jump[10];
   printf("Ecriture du programme dans la heap\n");
@@ -425,27 +445,18 @@ int main(int argc, char **argv)
                         return_stack[6],
                         return_stack[7]};
   if (write_memory(addr_puissance, jump_code, 10, process_pid, buffer_jump) ==
-      -1)
-  {
+      EXIT_FAILURE) {
     printf(
         "Erreur lors de l'appel de write_memory du programme dans la heap\n");
-    return -1;
+    return EXIT_FAILURE;
   }
 
-  // Traitement return_stack
-  // Get a pointeur in the heap
-
-  // Write the new function a this pointer (with write_memory)
-  // Translation of our function in bitecode
-  // Get size of this translation => put it in size_code
-
-  // trap_call_trap of this new function
-
-  // 4. Continuer l'exécution du processus
+  // Continuer l'éxécution du processus avec le nouveau code
   ptrace(PTRACE_CONT, process_pid, 0, 0);
 
-  // 5. Détacher du processus
+  // Détacher le processus
   ptrace(PTRACE_DETACH, process_pid, 0, 0);
 
   free(return_stack);
+  return EXIT_SUCCESS;
 }
